+++ Gate.h
@@ -1,21 +1,18 @@
 #ifndef GATE_H
 #define GATE_H
-
 #include<iostream>
 using namespace std;
-
 class Gate
 {
 private:
-    unsigned short loc1;
-    unsigned short loc2;
-    unsigned short loc3;
+    unsigned int loc1;
+    unsigned int loc2;
+    unsigned int loc3;
 public:
     Gate();
-    Gate(unsigned short l1, unsigned short l2, unsigned short l3);
-    unsigned short getLoc1();
-    unsigned short getLoc2();
-    unsigned short getLoc3();
-    
+    Gate(unsigned int l1, unsigned int l2, unsigned int l3);
+    unsigned int getLoc1();
+    unsigned int getLoc2();
+    unsigned int getLoc3();
 };
 #endif
